Refactored the expm() to sparse.linalg.matfuncs
